Refactor secondary app with domain/ui structure
