+++ app/domain/crypticSignalScheduler.js
