+++ bolt signal/app/domain/crypticSignalScheduler.js
