+++ bolt signal/app/domain/signalScheduler.js
